refactor : ErrorMessage 클래스명 변경 및 매직 레터럴 추가
