+++ src/test/java/com/flab/offcoupon/service/MemberServiceTest.java
@@ -11,7 +11,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static com.flab.offcoupon.exception.Constant.DUPLICATED_EMAIL;
+import static com.flab.offcoupon.exception.ErrorMessage.DUPLICATED_EMAIL;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.*;
